fix: app navigates to main page when token is expired
