+++ src/app/layout.tsx
@@ -1,15 +1,9 @@
 import type { Metadata } from "next";
 import { IBM_Plex_Mono } from "next/font/google";
 import "./globals.css";
-import { cookies } from "next/headers";
-import { getTokens } from "next-firebase-auth-edge";
-import {
-  clientConfig,
-  serverConfig,
-} from "@/services/next-firebase-auth-edge/config";
 import ToastProvider from "@/services/react-toastify/provider";
 import { AuthProvider } from "@/services/next-firebase-auth-edge/provider";
-import { toUser } from "@/services/next-firebase-auth-edge/utils";
+import { getUser } from "@/services/next-firebase-auth-edge/utils";
 import { ExpectedErrorBoundary } from "@/services/error-boundary/ErrorBoundary";
 import { LocaleProvider } from "@/services/locale/contex";
 
@@ -29,14 +23,7 @@ export default async function RootLayout({
 }: {
   children: React.ReactNode;
 }) {
-  const tokens = await getTokens(cookies(), {
-    apiKey: clientConfig.apiKey,
-    cookieName: serverConfig.cookieName,
-    cookieSignatureKeys: serverConfig.cookieSignatureKeys,
-    serviceAccount: serverConfig.serviceAccount,
-  });
-
-  const user = tokens ? toUser(tokens) : null;
+  const user = await getUser();
 
   return (
     <html lang="en">
